requirements.txt
fastapi
uvicorn
python-dotenv
pinecone-client
openai
pypdf
langchain
langchain_openai
langchain_community
python-multipart
python-dotenv